refactored code, reorganized lib, new test base
