+++ doc/BoardTest.html
@@ -4,7 +4,7 @@
 <head>
 <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
 
-<title>class ConnectFourTest - RDoc Documentation</title>
+<title>class BoardTest - RDoc Documentation</title>
 
 <link type="text/css" media="screen" href="./rdoc.css" rel="stylesheet">
 
@@ -49,7 +49,7 @@
     <nav id="file-list-section" class="section">
   <h3 class="section-header">Defined In</h3>
   <ul>
-    <li>test/connect_four_test.rb
+    <li>test/board_test.rb
   </ul>
 </nav>
 
@@ -169,7 +169,7 @@
 </nav>
 
 <div id="documentation">
-  <h1 class="class">class ConnectFourTest</h1>
+  <h1 class="class">class BoardTest</h1>
 
   <div id="description" class="description">
