+++ doc/doc/rdoc_css.html
@@ -59,6 +59,10 @@
   
     <li class="file"><a href="../Rakefile.html">Rakefile</a>
   
+    <li class="file"><a href="../doc/Board_html.html">Board.html</a>
+  
+    <li class="file"><a href="../doc/BoardTest_html.html">BoardTest.html</a>
+  
     <li class="file"><a href="../doc/ConnectFour_html.html">ConnectFour.html</a>
   
     <li class="file"><a href="../doc/ConnectFourSpec_html.html">ConnectFourSpec.html</a>
@@ -69,47 +73,23 @@
   
     <li class="file"><a href="../doc/Gemfile_lock_html.html">Gemfile_lock.html</a>
   
+    <li class="file"><a href="../doc/Player_html.html">Player.html</a>
+  
+    <li class="file"><a href="../doc/PlayerTest_html.html">PlayerTest.html</a>
+  
     <li class="file"><a href="../doc/README_rdoc_html.html">README_rdoc.html</a>
   
     <li class="file"><a href="../doc/Rakefile_html.html">Rakefile.html</a>
   
+    <li class="file"><a href="../doc/String_html.html">String.html</a>
+  
+    <li class="file"><a href="../doc/Tile_html.html">Tile.html</a>
+  
+    <li class="file"><a href="../doc/TileTest_html.html">TileTest.html</a>
+  
     <li class="file"><a href="../doc/created_rid.html">created.rid</a>
   
-    <li class="file"><a href="../doc/doc/ConnectFourSpec_html_html.html">ConnectFourSpec_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/ConnectFourTest_html_html.html">ConnectFourTest_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/ConnectFour_html_html.html">ConnectFour_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/Gemfile_html_html.html">Gemfile_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/Gemfile_lock_html_html.html">Gemfile_lock_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/README_rdoc_html_html.html">README_rdoc_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/Rakefile_html_html.html">Rakefile_html.html</a>
-  
     <li class="file"><a href="../doc/doc/created_rid_html.html">created_rid.html</a>
-  
-    <li class="file"><a href="../doc/doc/doc/created_rid_html_html.html">created_rid_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/index_html_html.html">index_html.html</a>
-  
-    <li class="file"><a href="../doc/doc/js/darkfish_js_html.html">darkfish_js.html</a>
-  
-    <li class="file"><a href="../doc/doc/js/jquery_js_html.html">jquery_js.html</a>
-  
-    <li class="file"><a href="../doc/doc/js/navigation_js_html.html">navigation_js.html</a>
-  
-    <li class="file"><a href="../doc/doc/js/search_index_js_html.html">search_index_js.html</a>
-  
-    <li class="file"><a href="../doc/doc/js/search_js_html.html">search_js.html</a>
-  
-    <li class="file"><a href="../doc/doc/js/searcher_js_html.html">searcher_js.html</a>
-  
-    <li class="file"><a href="../doc/doc/rdoc_css_html.html">rdoc_css.html</a>
-  
-    <li class="file"><a href="../doc/doc/table_of_contents_html_html.html">table_of_contents_html.html</a>
   
     <li class="file"><a href="../doc/index_html.html">index.html</a>
   
@@ -129,36 +109,6 @@
   
     <li class="file"><a href="../doc/table_of_contents_html.html">table_of_contents.html</a>
   
-    <li class="file"><a href="../lib/doc/ConnectFour_html.html">ConnectFour.html</a>
-  
-    <li class="file"><a href="../lib/doc/Grid_html.html">Grid.html</a>
-  
-    <li class="file"><a href="../lib/doc/Player_html.html">Player.html</a>
-  
-    <li class="file"><a href="../lib/doc/String_html.html">String.html</a>
-  
-    <li class="file"><a href="../lib/doc/created_rid.html">created.rid</a>
-  
-    <li class="file"><a href="../lib/doc/doc/created_rid_html.html">created_rid.html</a>
-  
-    <li class="file"><a href="../lib/doc/index_html.html">index.html</a>
-  
-    <li class="file"><a href="../lib/doc/js/darkfish_js.html">darkfish.js</a>
-  
-    <li class="file"><a href="../lib/doc/js/jquery_js.html">jquery.js</a>
-  
-    <li class="file"><a href="../lib/doc/js/navigation_js.html">navigation.js</a>
-  
-    <li class="file"><a href="../lib/doc/js/search_js.html">search.js</a>
-  
-    <li class="file"><a href="../lib/doc/js/search_index_js.html">search_index.js</a>
-  
-    <li class="file"><a href="../lib/doc/js/searcher_js.html">searcher.js</a>
-  
-    <li class="file"><a href="../lib/doc/rdoc_css.html">rdoc.css</a>
-  
-    <li class="file"><a href="../lib/doc/table_of_contents_html.html">table_of_contents.html</a>
-  
   </ul>
 </nav>
 
@@ -167,19 +117,25 @@
 
   <ul class="link-list">
   
+    <li><a href="../Board.html">Board</a>
+  
+    <li><a href="../BoardTest.html">BoardTest</a>
+  
     <li><a href="../ConnectFour.html">ConnectFour</a>
   
     <li><a href="../ConnectFourSpec.html">ConnectFourSpec</a>
   
     <li><a href="../ConnectFourTest.html">ConnectFourTest</a>
   
-    <li><a href="../Grid.html">Grid</a>
-  
     <li><a href="../Player.html">Player</a>
+  
+    <li><a href="../PlayerTest.html">PlayerTest</a>
   
     <li><a href="../String.html">String</a>
   
     <li><a href="../Tile.html">Tile</a>
+  
+    <li><a href="../TileTest.html">TileTest</a>
   
   </ul>
 </nav>
