+++ doc/js/search_index.js
@@ -1,1 +1,1 @@
-var search_data = {"index":{"searchIndex":["connectfour","connectfourspec","connectfourtest","grid","player","string","tile","ask_player_if_he_wants_to_play_again()","assign_to()","build()","checks_if_column_is_in_grid?()","colorize()","connect_disc_with_grid()","draw_grid()","drop_disc()","finds_diagonal_connections?()","finds_horizontal_connections?()","finds_vertical_connections?()","info_range_error()","informs_player_about_full_column()","is_connected?()","matches_pattern?()","new()","new()","new()","new()","next_turn()","play()","play()","play_again()","play_turn()","prompt()","provide_column()","quit()","reset()","show_rules()","toggle_player()","wins()","gemfile","gemfile.lock","readme","rakefile","connectfour.html","connectfourspec.html","connectfourtest.html","gemfile.html","gemfile_lock.html","readme_rdoc.html","rakefile.html","created.rid","connectfourspec_html.html","connectfourtest_html.html","connectfour_html.html","gemfile_html.html","gemfile_lock_html.html","readme_rdoc_html.html","rakefile_html.html","created_rid.html","created_rid_html.html","index_html.html","darkfish_js.html","jquery_js.html","navigation_js.html","search_index_js.html","search_js.html","searcher_js.html","rdoc_css.html","table_of_contents_html.html","index.html","darkfish.js","jquery.js","navigation.js","search.js","search_index.js","searcher.js","rdoc.css","table_of_contents.html","connectfour.html","grid.html","player.html","string.html","created.rid","created_rid.html","index.html","darkfish.js","jquery.js","navigation.js","search.js","search_index.js","searcher.js","rdoc.css","table_of_contents.html"],"longSearchIndex":["connectfour","connectfourspec","connectfourtest","grid","player","string","tile","connectfour#ask_player_if_he_wants_to_play_again()","tile#assign_to()","grid#build()","connectfour#checks_if_column_is_in_grid?()","string#colorize()","grid#connect_disc_with_grid()","grid#draw_grid()","grid#drop_disc()","grid#finds_diagonal_connections?()","grid#finds_horizontal_connections?()","grid#finds_vertical_connections?()","connectfour#info_range_error()","grid#informs_player_about_full_column()","grid#is_connected?()","grid#matches_pattern?()","connectfour::new()","grid::new()","player::new()","tile::new()","connectfour#next_turn()","connectfour#play()","player#play()","connectfour#play_again()","connectfour#play_turn()","string#prompt()","grid#provide_column()","connectfour#quit()","grid#reset()","connectfour#show_rules()","connectfour#toggle_player()","player#wins()","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","","",""],"info":[["ConnectFour","","ConnectFour.html","","<p>CONNECT FOUR CLASSES\n"],["ConnectFourSpec","","ConnectFourSpec.html","",""],["ConnectFourTest","","ConnectFourTest.html","",""],["Grid","","Grid.html","",""],["Player","","Player.html","",""],["String","","String.html","","<p>HELPERS\n"],["Tile","","Tile.html","",""],["ask_player_if_he_wants_to_play_again","ConnectFour","ConnectFour.html#method-i-ask_player_if_he_wants_to_play_again","()",""],["assign_to","Tile","Tile.html#method-i-assign_to","(player)",""],["build","Grid","Grid.html#method-i-build","()",""],["checks_if_column_is_in_grid?","ConnectFour","ConnectFour.html#method-i-checks_if_column_is_in_grid-3F","(column)",""],["colorize","String","String.html#method-i-colorize","(color_code)",""],["connect_disc_with_grid","Grid","Grid.html#method-i-connect_disc_with_grid","(column, player)",""],["draw_grid","Grid","Grid.html#method-i-draw_grid","()",""],["drop_disc","Grid","Grid.html#method-i-drop_disc","(column, player)",""],["finds_diagonal_connections?","Grid","Grid.html#method-i-finds_diagonal_connections-3F","(is_anti, column, row, player)",""],["finds_horizontal_connections?","Grid","Grid.html#method-i-finds_horizontal_connections-3F","(row, player)",""],["finds_vertical_connections?","Grid","Grid.html#method-i-finds_vertical_connections-3F","(column, player)",""],["info_range_error","ConnectFour","ConnectFour.html#method-i-info_range_error","()",""],["informs_player_about_full_column","Grid","Grid.html#method-i-informs_player_about_full_column","(player,column)",""],["is_connected?","Grid","Grid.html#method-i-is_connected-3F","(col, row, player)",""],["matches_pattern?","Grid","Grid.html#method-i-matches_pattern-3F","(player, result)",""],["new","ConnectFour","ConnectFour.html#method-c-new","(number_of_columns, number_of_rows, number_of_discs_to_connect)",""],["new","Grid","Grid.html#method-c-new","(number_of_columns, number_of_rows, number_of_discs_to_connect)",""],["new","Player","Player.html#method-c-new","(color, foreground_colorcode, background_colorcode)",""],["new","Tile","Tile.html#method-c-new","(col,row)",""],["next_turn","ConnectFour","ConnectFour.html#method-i-next_turn","()",""],["play","ConnectFour","ConnectFour.html#method-i-play","()",""],["play","Player","Player.html#method-i-play","(column, grid)",""],["play_again","ConnectFour","ConnectFour.html#method-i-play_again","()",""],["play_turn","ConnectFour","ConnectFour.html#method-i-play_turn","(player_drops_disc_to_column)",""],["prompt","String","String.html#method-i-prompt","()",""],["provide_column","Grid","Grid.html#method-i-provide_column","(column)",""],["quit","ConnectFour","ConnectFour.html#method-i-quit","()",""],["reset","Grid","Grid.html#method-i-reset","()",""],["show_rules","ConnectFour","ConnectFour.html#method-i-show_rules","()",""],["toggle_player","ConnectFour","ConnectFour.html#method-i-toggle_player","()",""],["wins","Player","Player.html#method-i-wins","()",""],["Gemfile","","Gemfile.html","","<p>source :rubygems\n<p>gem &#39;rake&#39; gem &#39;minitest&#39;, &#39;~&gt; 5.0.8&#39;\n"],["Gemfile.lock","","Gemfile_lock.html","","<p>GEM\n\n<pre>remote: http://rubygems.org/\nspecs:\n  minitest (5.0.8)\n  rake (10.1.0)</pre>\n<p>PLATFORMS\n"],["README","","README_rdoc.html","","<p>Connect four\n<p>Console based “Vier gewinnt”.\n<p>Game play\n"],["Rakefile","","Rakefile.html","","<p>require &#39;rubygems&#39; require &#39;bundler/setup&#39; require\n&#39;rake/testtask&#39;\n<p>task :default …\n"],["ConnectFour.html","","doc/ConnectFour_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFourSpec.html","","doc/ConnectFourSpec_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFourTest.html","","doc/ConnectFourTest_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Gemfile.html","","doc/Gemfile_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Gemfile_lock.html","","doc/Gemfile_lock_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["README_rdoc.html","","doc/README_rdoc_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Rakefile.html","","doc/Rakefile_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["created.rid","","doc/created_rid.html","","<p>Fri, 25 Oct 2013 19:50:33 +0200 ./doc/ConnectFour.html  Wed, 23 Oct 2013\n01:44:30 +0200 ./doc/ConnectFourSpec.html …\n"],["ConnectFourSpec_html.html","","doc/doc/ConnectFourSpec_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFourTest_html.html","","doc/doc/ConnectFourTest_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFour_html.html","","doc/doc/ConnectFour_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Gemfile_html.html","","doc/doc/Gemfile_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Gemfile_lock_html.html","","doc/doc/Gemfile_lock_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["README_rdoc_html.html","","doc/doc/README_rdoc_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Rakefile_html.html","","doc/doc/Rakefile_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["created_rid.html","","doc/doc/created_rid_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["created_rid_html.html","","doc/doc/doc/created_rid_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["index_html.html","","doc/doc/index_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["darkfish_js.html","","doc/doc/js/darkfish_js_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["jquery_js.html","","doc/doc/js/jquery_js_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["navigation_js.html","","doc/doc/js/navigation_js_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["search_index_js.html","","doc/doc/js/search_index_js_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["search_js.html","","doc/doc/js/search_js_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["searcher_js.html","","doc/doc/js/searcher_js_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["rdoc_css.html","","doc/doc/rdoc_css_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["table_of_contents_html.html","","doc/doc/table_of_contents_html_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["index.html","","doc/index_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["darkfish.js","","doc/js/darkfish_js.html","","\n<pre>Darkfish Page Functions\n$Id: darkfish.js 53 2009-01-07 02:52:03Z deveiant $\n\nAuthor: Michael Granger ...</pre>\n"],["jquery.js","","doc/js/jquery_js.html","","\n<pre>!\n jQuery JavaScript Library v1.6.2\n http://jquery.com/\n\n Copyright 2011, John Resig\n Dual licensed under ...</pre>\n"],["navigation.js","","doc/js/navigation_js.html","","\n<pre>Navigation allows movement using the arrow keys through the search results.\n\nWhen using this library ...</pre>\n"],["search.js","","doc/js/search_js.html","","<p>Search = function(data, input, result) {\n\n<pre>this.data = data;\nthis.$input = $(input);\nthis.$result = $(result); ...</pre>\n"],["search_index.js","","doc/js/search_index_js.html","","<p>var search_data = {“index”:{“searchIndex”:,“longSearchIndex”:,“info”:[,[“ …\n"],["searcher.js","","doc/js/searcher_js.html","","<p>Searcher = function(data) {\n\n<pre>this.data = data;\nthis.handlers = [];</pre>\n<p>}\n"],["rdoc.css","","doc/rdoc_css.html","","\n<pre>&quot;Darkfish&quot; Rdoc CSS\n$Id: rdoc.css 54 2009-01-27 01:09:48Z deveiant $\n\nAuthor: Michael Granger &lt;ged@FaerieMUD.org&gt; ...</pre>\n"],["table_of_contents.html","","doc/table_of_contents_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFour.html","","lib/doc/ConnectFour_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Grid.html","","lib/doc/Grid_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Player.html","","lib/doc/Player_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["String.html","","lib/doc/String_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["created.rid","","lib/doc/created_rid.html","","<p>Fri, 25 Oct 2013 22:58:19 +0200 ./connect_four.rb       Fri, 25 Oct 2013\n22:54:14 +0200 ./doc/created.rid …\n"],["created_rid.html","","lib/doc/doc/created_rid_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["index.html","","lib/doc/index_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["darkfish.js","","lib/doc/js/darkfish_js.html","","\n<pre>Darkfish Page Functions\n$Id: darkfish.js 53 2009-01-07 02:52:03Z deveiant $\n\nAuthor: Michael Granger ...</pre>\n"],["jquery.js","","lib/doc/js/jquery_js.html","","\n<pre>!\n jQuery JavaScript Library v1.6.2\n http://jquery.com/\n\n Copyright 2011, John Resig\n Dual licensed under ...</pre>\n"],["navigation.js","","lib/doc/js/navigation_js.html","","\n<pre>Navigation allows movement using the arrow keys through the search results.\n\nWhen using this library ...</pre>\n"],["search.js","","lib/doc/js/search_js.html","","<p>Search = function(data, input, result) {\n\n<pre>this.data = data;\nthis.$input = $(input);\nthis.$result = $(result); ...</pre>\n"],["search_index.js","","lib/doc/js/search_index_js.html","","<p>var search_data =\n{“index”:{“searchIndex”:,“longSearchIndex”:,“info”:[[“ConnectFour”,“”,“ConnectFour.html”,“”,“&lt;p&gt;CONNECT\n…\n"],["searcher.js","","lib/doc/js/searcher_js.html","","<p>Searcher = function(data) {\n\n<pre>this.data = data;\nthis.handlers = [];</pre>\n<p>}\n"],["rdoc.css","","lib/doc/rdoc_css.html","","\n<pre>&quot;Darkfish&quot; Rdoc CSS\n$Id: rdoc.css 54 2009-01-27 01:09:48Z deveiant $\n\nAuthor: Michael Granger &lt;ged@FaerieMUD.org&gt; ...</pre>\n"],["table_of_contents.html","","lib/doc/table_of_contents_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"]]}}
+var search_data = {"index":{"searchIndex":["board","boardtest","connectfour","connectfourspec","connectfourtest","player","playertest","string","tile","tiletest","ask_player_if_he_wants_to_play_again()","ask_player_to_drop_disc_in_a_column()","assign_to()","build_tilemap()","can_play_in_column?()","check_if_player_wins()","check_that_column_is_in_range()","checks_for_diagonal_tile_at_position()","colorize()","column_is_full?()","draw_board()","drop_disc_to()","drops_disc_to_column_in_board()","finds_diagonal_connections?()","finds_horizontal_connections?()","finds_vertical_connections?()","get_diagonal_search_range()","get_diagonal_tile_at_position()","get_highest_number_of_discs_to_connect()","info_player_about_full_column()","info_range_error()","insert_disc_to()","is_connected?()","is_in_range?()","matches_connection_in_search_result?()","new()","new()","new()","new()","next_turn()","play()","play_again()","play_turn()","prompt()","provide_column_for_discs()","quit()","reset_tilemap()","show_rules()","toggle_player()","wins()","gemfile","gemfile.lock","readme","rakefile","board.html","boardtest.html","connectfour.html","connectfourspec.html","connectfourtest.html","gemfile.html","gemfile_lock.html","player.html","playertest.html","readme_rdoc.html","rakefile.html","string.html","tile.html","tiletest.html","created.rid","created_rid.html","index.html","darkfish.js","jquery.js","navigation.js","search.js","search_index.js","searcher.js","rdoc.css","table_of_contents.html"],"longSearchIndex":["board","boardtest","connectfour","connectfourspec","connectfourtest","player","playertest","string","tile","tiletest","connectfour#ask_player_if_he_wants_to_play_again()","connectfour#ask_player_to_drop_disc_in_a_column()","tile#assign_to()","board#build_tilemap()","board#can_play_in_column?()","board#check_if_player_wins()","connectfour#check_that_column_is_in_range()","board#checks_for_diagonal_tile_at_position()","string#colorize()","board#column_is_full?()","board#draw_board()","board#drop_disc_to()","player#drops_disc_to_column_in_board()","board#finds_diagonal_connections?()","board#finds_horizontal_connections?()","board#finds_vertical_connections?()","board#get_diagonal_search_range()","board#get_diagonal_tile_at_position()","board#get_highest_number_of_discs_to_connect()","board#info_player_about_full_column()","connectfour#info_range_error()","board#insert_disc_to()","board#is_connected?()","connectfour#is_in_range?()","board#matches_connection_in_search_result?()","board::new()","connectfour::new()","player::new()","tile::new()","connectfour#next_turn()","connectfour#play()","connectfour#play_again()","connectfour#play_turn()","string#prompt()","board#provide_column_for_discs()","connectfour#quit()","board#reset_tilemap()","connectfour#show_rules()","connectfour#toggle_player()","player#wins()","","","","","","","","","","","","","","","","","","","","","","","","","","","","",""],"info":[["Board","","Board.html","",""],["BoardTest","","BoardTest.html","",""],["ConnectFour","","ConnectFour.html","","<p>CONNECT FOUR\n"],["ConnectFourSpec","","ConnectFourSpec.html","",""],["ConnectFourTest","","ConnectFourTest.html","",""],["Player","","Player.html","",""],["PlayerTest","","PlayerTest.html","",""],["String","","String.html","","<p>HELPERS\n"],["Tile","","Tile.html","",""],["TileTest","","TileTest.html","",""],["ask_player_if_he_wants_to_play_again","ConnectFour","ConnectFour.html#method-i-ask_player_if_he_wants_to_play_again","()",""],["ask_player_to_drop_disc_in_a_column","ConnectFour","ConnectFour.html#method-i-ask_player_to_drop_disc_in_a_column","()",""],["assign_to","Tile","Tile.html#method-i-assign_to","(player)",""],["build_tilemap","Board","Board.html#method-i-build_tilemap","()",""],["can_play_in_column?","Board","Board.html#method-i-can_play_in_column-3F","(column, player)",""],["check_if_player_wins","Board","Board.html#method-i-check_if_player_wins","(column, row, player)",""],["check_that_column_is_in_range","ConnectFour","ConnectFour.html#method-i-check_that_column_is_in_range","(column)",""],["checks_for_diagonal_tile_at_position","Board","Board.html#method-i-checks_for_diagonal_tile_at_position","(column, row, diagonal_direction, position)",""],["colorize","String","String.html#method-i-colorize","(color_code)",""],["column_is_full?","Board","Board.html#method-i-column_is_full-3F","(column)",""],["draw_board","Board","Board.html#method-i-draw_board","()",""],["drop_disc_to","Board","Board.html#method-i-drop_disc_to","(column, player)",""],["drops_disc_to_column_in_board","Player","Player.html#method-i-drops_disc_to_column_in_board","(column, board)",""],["finds_diagonal_connections?","Board","Board.html#method-i-finds_diagonal_connections-3F","(is_antidiagonal, column, row, player)",""],["finds_horizontal_connections?","Board","Board.html#method-i-finds_horizontal_connections-3F","(row, player)",""],["finds_vertical_connections?","Board","Board.html#method-i-finds_vertical_connections-3F","(column, player)",""],["get_diagonal_search_range","Board","Board.html#method-i-get_diagonal_search_range","()",""],["get_diagonal_tile_at_position","Board","Board.html#method-i-get_diagonal_tile_at_position","(column, row, diagonal_direction, position)",""],["get_highest_number_of_discs_to_connect","Board","Board.html#method-i-get_highest_number_of_discs_to_connect","(number_of_discs)",""],["info_player_about_full_column","Board","Board.html#method-i-info_player_about_full_column","(player,column)",""],["info_range_error","ConnectFour","ConnectFour.html#method-i-info_range_error","()",""],["insert_disc_to","Board","Board.html#method-i-insert_disc_to","(column, player)",""],["is_connected?","Board","Board.html#method-i-is_connected-3F","(col, row, player)",""],["is_in_range?","ConnectFour","ConnectFour.html#method-i-is_in_range-3F","(column)",""],["matches_connection_in_search_result?","Board","Board.html#method-i-matches_connection_in_search_result-3F","(player, search_result)",""],["new","Board","Board.html#method-c-new","(number_of_columns, number_of_rows, number_of_discs_to_connect)",""],["new","ConnectFour","ConnectFour.html#method-c-new","(number_of_columns, number_of_rows, number_of_discs_to_connect)",""],["new","Player","Player.html#method-c-new","(color, foreground_colorcode, background_colorcode)",""],["new","Tile","Tile.html#method-c-new","(column,row)",""],["next_turn","ConnectFour","ConnectFour.html#method-i-next_turn","()",""],["play","ConnectFour","ConnectFour.html#method-i-play","()",""],["play_again","ConnectFour","ConnectFour.html#method-i-play_again","()",""],["play_turn","ConnectFour","ConnectFour.html#method-i-play_turn","(column)",""],["prompt","String","String.html#method-i-prompt","()",""],["provide_column_for_discs","Board","Board.html#method-i-provide_column_for_discs","(column)",""],["quit","ConnectFour","ConnectFour.html#method-i-quit","()",""],["reset_tilemap","Board","Board.html#method-i-reset_tilemap","()",""],["show_rules","ConnectFour","ConnectFour.html#method-i-show_rules","()",""],["toggle_player","ConnectFour","ConnectFour.html#method-i-toggle_player","()",""],["wins","Player","Player.html#method-i-wins","()",""],["Gemfile","","Gemfile.html","","<p>source :rubygems\n<p>gem &#39;rake&#39; gem &#39;minitest&#39;, &#39;~&gt; 5.0.8&#39;\n"],["Gemfile.lock","","Gemfile_lock.html","","<p>GEM\n\n<pre>remote: http://rubygems.org/\nspecs:\n  minitest (5.0.8)\n  rake (10.1.0)</pre>\n<p>PLATFORMS\n"],["README","","README_rdoc.html","","<p>Connect four\n<p>Console based “Vier gewinnt”.\n<p>Game play\n"],["Rakefile","","Rakefile.html","","<p>require &#39;rubygems&#39; require &#39;bundler/setup&#39; require\n&#39;rake/testtask&#39;\n<p>task :default …\n"],["Board.html","","doc/Board_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["BoardTest.html","","doc/BoardTest_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFour.html","","doc/ConnectFour_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFourSpec.html","","doc/ConnectFourSpec_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["ConnectFourTest.html","","doc/ConnectFourTest_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Gemfile.html","","doc/Gemfile_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Gemfile_lock.html","","doc/Gemfile_lock_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Player.html","","doc/Player_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["PlayerTest.html","","doc/PlayerTest_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["README_rdoc.html","","doc/README_rdoc_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Rakefile.html","","doc/Rakefile_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["String.html","","doc/String_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["Tile.html","","doc/Tile_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["TileTest.html","","doc/TileTest_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["created.rid","","doc/created_rid.html","","<p>Mon, 28 Oct 2013 00:37:52 +0100 ./app.rb        Mon, 28 Oct 2013 00:20:05\n+0100 ./doc/created.rid    …\n"],["created_rid.html","","doc/doc/created_rid_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["index.html","","doc/index_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"],["darkfish.js","","doc/js/darkfish_js.html","","\n<pre>Darkfish Page Functions\n$Id: darkfish.js 53 2009-01-07 02:52:03Z deveiant $\n\nAuthor: Michael Granger ...</pre>\n"],["jquery.js","","doc/js/jquery_js.html","","\n<pre>!\n jQuery JavaScript Library v1.6.2\n http://jquery.com/\n\n Copyright 2011, John Resig\n Dual licensed under ...</pre>\n"],["navigation.js","","doc/js/navigation_js.html","","\n<pre>Navigation allows movement using the arrow keys through the search results.\n\nWhen using this library ...</pre>\n"],["search.js","","doc/js/search_js.html","","<p>Search = function(data, input, result) {\n\n<pre>this.data = data;\nthis.$input = $(input);\nthis.$result = $(result); ...</pre>\n"],["search_index.js","","doc/js/search_index_js.html","","<p>var search_data =\n{“index”:{“searchIndex”:,“longSearchIndex”:,“info”:[,[“BoardTest”,“”,“BoardTest.html”,“”,“”],[“ConnectFour”,“”,“ConnectFour.html”,“”,“&lt;p&gt;CONNECT\n…\n"],["searcher.js","","doc/js/searcher_js.html","","<p>Searcher = function(data) {\n\n<pre>this.data = data;\nthis.handlers = [];</pre>\n<p>}\n"],["rdoc.css","","doc/rdoc_css.html","","\n<pre>&quot;Darkfish&quot; Rdoc CSS\n$Id: rdoc.css 54 2009-01-27 01:09:48Z deveiant $\n\nAuthor: Michael Granger &lt;ged@FaerieMUD.org&gt; ...</pre>\n"],["table_of_contents.html","","doc/table_of_contents_html.html","","<p>&lt;!DOCTYPE html&gt;\n<p>&lt;html&gt; &lt;head&gt; &lt;meta content=“text/html; charset=UTF-8” …\n"]]}}
