+++ test/player_test.rb
@@ -1,6 +1,6 @@
 require 'connect_four_spec_test'
 
-class ConnectFourTest < ConnectFourSpec
+class PlayerTest < ConnectFourSpec
 	#def test_hello_world
 		#assert_equal "hello world", ConnectFour.hello_world
 	#end
